fix: binding property of textarea and input
